Make IOccurrenceStatsDAO
- This DAO can work with IMentionables to store retrieve and perform simple stats
    queries on them

- Make the emoji and the chat entity dao delegate to it

+++ compute/src/main/java/com/chatalytics/compute/storm/bolt/EntityExtractionBolt.java
@@ -124,7 +124,7 @@ public class EntityExtractionBolt extends BaseRichBolt {
 
                 if (!tag.isEndTag) {
                     ChatEntity existingEntity = entities.remove(entity);
-                    long occurrences;
+                    int occurrences;
                     if (existingEntity == null) {
                         occurrences = 1;
                     } else {
